requirements.txt
apify-client==1.6.1
asyncpraw==7.7.1
bittensor==6.9.3
jupyter==1.0.0
numpy==1.26.2
# Bittensor require pydantic < 2.0.0
pydantic==1.10.0
python-dotenv==1.0.0
pytz==2023.3.post1
rich==13.7.0
torch==2.1.2
wandb==0.16.2
pymongo==4.6.1